interpreter/requirements.txt
waitress~=2.0.0
flask~=2.0.1
flask-cors
pydub~=0.25.1
python-jose
tensorflow-cpu
librosa~=0.8.1
pytz~=2021.1
passlib~=1.7.4
bson~=0.5.10
pymongo~=3.12.0
matplotlib~=3.4.2
tqdm~=4.62.0
requests~=2.26.0
numpy~=1.19.5
scipy~=1.7.1
Werkzeug~=2.0.1